Create README.md
https://www.kaggle.com/c/titanic/overview

How a Data Scientist finds a solution
Titanic survival prediction is a classical problem, predict the outcome of a binary event. In laymen terms this means, it either occurred or did not occur. For example, you won or did not win, you passed the test or did not pass the test, you were accepted or not accepted, and you get the point. Some common business application are  healthcare's mortality rate or survival analysis. 

Binary events create an interesting dynamic, because we know statistically, a random guess should achieve a 50% accuracy rate, without creating one single algorithm or writing one single line of code.